LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/inline/set-shorthand-expected.txt

CONSOLE MESSAGE: TypeError: CSS.deg is not a function. (In 'CSS.deg(0)', 'CSS.deg' is undefined)

Harness Error (FAIL), message = TypeError: CSS.deg is not a function. (In 'CSS.deg(0)', 'CSS.deg' is undefined)
